refactor: 도메인 타입 entities/model 로 통합
- src/types/types.ts 의 도메인 타입을 슬라이스별 model 로 이전
  - Cafe/Menu/Tag/Country/Flavors/CoffeeBean → entities/cafe/model/types
  - BookmarkCafe → entities/bookmark/model/types
- @/types/types import 11곳 새 경로로 정정, 옛 src/types 삭제
- cafe/model 의 자기참조 import 제거 (프리미티브가 동일 파일로 병합)
