+++ src/entities/bookmark/model/types.ts
@@ -1,0 +1,7 @@
+export interface BookmarkCafe {
+  id: string;
+  name: string;
+  mainImageUrl: string[];
+  location: string;
+  isBookmarked?: boolean;
+}
